feat: Have display show clicked buttons/operators
Wired up the number and operator buttons to populate the display. Next
challenge is going to be figuring out how to store the numbers and
operators that are clicked. Thinking an array, with num1 and num2 in
spots 0 and 2, and the operator in spot 1. But then how do you tell it
to move on to the next array spot?