feat: Handle no num1, operator, or num2
Handles cases when the user does not input one of the necessary
arguments for the getAnswer function.

Also, changed the initial values of num1 and num2 from 0 back to empty
strings. Trying to reset them in the numBtns event listener meant that
the values kept clearing on each click. So only the last number would be
saved in the variable.

Next step: wire up the clear button.