feat: keyboard input for nums and decimal
Changed the inside of the numBtns event listener to a function for
reuse. Figured out how to disable the keydown decimal button by just
appending an empty string to the display.

Next: Do the same thing to the operators. Need to probably make a
function to handle the multiply and divide characters.