feat: add keyboard input functionality
Tweaked some of the event listeners and make their content into
functions for reuse with keyboard input.

Also, running into an issue with buttons keeping the focus after they've
been clicked, when moving between mouse and keyboard input. I currently
have preventDefault on the enter button to fix this, but I want to try
using `mousedown` instead of `click` on the events to remove the focus
outline on the buttons.