Fix new calc after equals bug
A bug popped up where, since I was storing num1, starting a new
calculation after pressing equals would just append the new number to
num1 instead of starting over with num1 as empty. Added a lastEqualsFlag
to take of that. Could probably streamline it a bit.

Also changed the appearance of the calculator.